requirements.txt
alembic==1.14.0
asyncpg==0.30.0
fastapi==0.115.6
flake8==7.1.1
greenlet==3.1.1
gunicorn==23.0.0
httpx==0.28.1
isort==5.13.2
pydantic==2.10.3
pydantic-settings==2.7.0
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.0
SQLAlchemy==2.0.36
uvicorn==0.32.1
websockets==14.1
